python/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
MouseInfo==0.1.3
Pillow==9.1.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pyserial==3.5
pytweening==1.0.4
requests==2.27.1
six==1.16.0
spotipy==2.19.0
Unidecode==1.3.4
urllib3==1.26.9
